+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ## Added
 
-- [TD-2048] New column system_required in Rule
+- [TD-2048] New Rule Type system param ´system_required´ of type boolean with hidden property
 
 ## Changed
 
-- [TD-2048] Optional system in Rule Implementations
+- [TD-2048] Attribute system in Rule Implementation requirement is defined in Rule Type system params as ´system_required´
 
 ## [3.7.0] 2019-09-30